+++ OpenEmuBase/TPCircularBuffer.c
@@ -25,7 +25,7 @@ bool TPCircularBufferInit(TPCircularBuffer *buffer, int length) {
     int retries = 3;
     while ( true ) {
 
-        buffer->length = round_page(length);    // We need whole page sizes
+        buffer->length = (int)round_page(length);    // We need whole page sizes
 
         // Temporarily allocate twice the length, so we have the contiguous address space to
         // support a second instance of the buffer directly after
